Remove idle timeout key (not used and removed)
The key idle timeout was removed in 10.10. Furthermore, it is the job's responsibility to terminate itself when it has idled for a specific time. This was seemingly only a key to send an idle timeout value to the job itself. launchd won't kill the job due to any idle timeout settings.