zfs_append_partition() was broken when using by-id
Even when whole_disk=1, the displayed InvariantDisks by-id device names
are already pointing at the partition not the whole device. This means
that zfs_append_partition() has nothing to do, and should just return
the length.

This is not the case for by-serial and by-path device names when
whole_disk=1 because in those cases the displayed device names point to
the whole disks, and so need tweaking in order to get the name of the
partition actually containing the vdev.

The consequence of the bug in the by-id case was that users would get
the error "no such device in pool" whenever they tried to use zpool
replace, zpool attach, or zpool online with whole_disk=1 vdevs.

Closes #457