DLPX-42518 ZFS still fails to send some holes
Because of the way the if statements in traverse_visitbp are structured,
If a hole has birth time 0 in an incremental send that contains enough
objects that an object could have been reallocated, then the hole would
be skipped. The fix is to move the clause testing for the possibility of
reallocation to the nested if statement, rather than the if() {} else if
() {} block. We also fix the unit test by asserting its result.