Fix defer_destroy regression
During 716d05de, most, but not all, of zfsonlinux/zfs@962d5242 and
zfsonlinux/zfs@09c0b8fe were reverted.  For now, the simplest fix is to
remove the remaining code that wasn't reverted, and wait for those
commits to be upstreamed.

Another option would be to cherry-pick those commits back in now, and
put up with the ongoing merge conflict unless they get upstreamed.