6922 Emit ESC_ZFS_VDEV_REMOVE_AUX after removing an aux device
Reviewed by: George Wilson <george.wilson@delphix.com>
Reviewed by: Matthew Ahrens <mahrens@delphix.com>

Always do a config sync after removing an aux (spare, zil, or l2arc) device.
AFAICT this wasn't being done because it is slow and was deemed unnecessary.
However, it should be such a rare operation that speed doesn't matter, and not
doing it results in two problems:

1) It is theoretically possible to remove an aux device from one pool and
   attach it to another, then lose power.  When power is restored, both
   pools would think that they own the aux device.
2) Removal of the aux device doesn't send any useful sysevents to
   userland.

Porting notes: events renamed to fit our style, FM_EREPORT_ZFS_ ...