store insert_time, tie-break m==m with size
use insert_time to bail early from the move func if the
insertion was too recent, since we mostly care about old
entries.

(this needs a bit of watching, since multiple LATER
returns will turn into a NO, and we will not be asked
again).

(could store the count of LATERs in the avl with an update)