Improve ABD performance on openzfsonosx port
O3X's vmem layer is a client of one of xnu's kernel memory
allocators, and has to accomplish a balancing act between:

a) not talking too often to xnu, because xnu's allocator
will stall the whole system dealing with large numbers of
simultaneous requests (vmem and kmem do heavy concurrency
on multiprocessor systems) and because xnu's freelist
handling degrades badly when dealing with many small
allocations and frees

b) reasonable overall memory (and CPU) use for ZFS

(There is a secondary goal (c) namely visibility into the
memory system via kstats).

Before ABD the tension between the two was driven by the
observation that the strongest memory use was compressed
ARC buffers, which come in discrete sizes due to the
minimum SPA block size and larger ashifts on vdevs.
We segregated the allocations by size as much as possible
and found a good balance between (a) and (b).

ABD removes the segregation of compressed ARC buffers by
scattering them into 1k chunks in their own arena.

This broke the balance between (a) and (b), leading to a
huge ballooning of fragmentation, leading to either or
both of huge overall memory consumption or a tiny ARC.

This patch recovers a decent balance between (a) and (b)
by (1) adjusting the vmem and kmem plumbing of the tree of
arenas under the abd_chunk cache and the other large and
busy kmem caches and (2) introducing a system that
attempts to defragment the quantum caches and vmem bucket
when it eventually appears (e.g. after a large eviction
after an ARC has accumulated on a multi-processor system
doing multi-threaded I/O, which leads to maximal
scattering of large ARC buffers into multiple segments and
spans shared with smaller buffers with longer lifespans).

The moving system spawns a thread in arc.c that
periodically checks for fragmentation and, finding it,
subject to timeout, walks the ARC multilists from their
heads looking for candidate ABDs to move.  We lock ABD
access with a new per-ABD mutex, and move the locked data
into newly allocated memory.  This "unsticks" old abd
chunks that "kidnap" spans that are otherwise freeable and
unlikely to be immediately allocated into by new ABD
allocations.  We also keep a per-ABD hrtime value so we do
not move ephemeral ABDs or immediately re-move ABDs which
have been recently moved.

For visibility several new abdstat and arcstat kstats have
been made.

arc_meta_min has also been reduced.  At pool import time
with ashift=9 and multiple vdevs and datasets, a large
amount of 512-byte metadata is pulled into abd (which uses
1k chunks).  Especially where mds and mdworker operate on
the datasets immediately after they are mounted, this
"early' metadata is mixed in with larger "early" file data
block allocation.  When the file data blocks are evicted,
many kmem_cache spans remain "kidnapped" by the small
metadata which, given a large arc_meta_min stays put
perhaps for the whole system uptime, even if the metaddata
is not reused at all.  With a smaller arc_meta_min,
low-value at-import metadata does not linger
counterproductively, but high-value metadata at import and
since import still gets a good share of the overall ARC
memory (this can be seen in the kstats) and is protected
from extreme ARC shrinks.

As arc_meta_min is runtime tunable, users with unusually
heavy "late" metadata activity can boost arc_meta_min to
suit their workloads.  However we should not hurt all
users principally for the benefit users who (as concrete
examples of high-metadata workload) turn on deduplication
or who set both a large count property and a small
recordsize/volblocksize property

Additionally:

* fixes for DEBUG userland and kernel building and operation