Tune abd.c : new vmem arena, extra kstats
abd's chunk cache was parented to the zio metadata
arena (zio_arena), which was suboptimal in several ways:

-- the chunks hold a mix of metadata and file data, so
   visibility into metadata v file data split was lost

--  all chunks are the same size:
    -- zio_arena does qcaching, which is unnecessary ove
       overhead both in terms of large import spans and in
       terms of work per abd chunk allocation

Instead, we use an arena that imports 64k spans which
appears to be a decent trade-off between visibility of the
vmem kstats, a large enough import span size so that the
contention in bucket_heap and vba-and-below is not
dramatically increased, and reducing the number of slabs
exposed to "kidnap" by lingering small numbers of abd
chunks.

abd's kmem_cache also forbade debugging in two ways:
firstly by setting a 0 alignment and secondly by
explicitly (and redundantly) setting KMC_NOTOUCH.
According to the comment in the openzfs code, this is a
policy decision relating to kernel core dumping, and we do
that differently in xnu anyway.

Turn on some kmem debugging on the abd chunk cache.

New abdstats to track zfs metadata vs file data in the abd
chunk cache.

Count the number of abd_alloc() calls where size < chunk size,
but do not actually linearize them while chunk size is 1k,
since the maximum waste will be only 512 bytes and
scatter_chunk_waste is only tens of MiB.

If chunk size is raised to PAGESIZE, for example, then
linearizing smaller abd_alloc()s will make more sense;
these smaller blocks would rely on the different sizes of
zio caches to reduce internal fragmentation
instead (although some extra slabs could stick around in
each kmem_cache, plus in the underlying qcache in o3x's
implementation).