use a much smaller default arc_meta_min
At pool import time, where ashift=9 and multiple vdevs and
datsets are involved, a large amount of 512-byte metadata
is pulled into abd.

And, for a machine with sufficient memory (~ 8GiB or more)
such that arc_c_min is large (and arc_meta_min = arc_c_min
/ 2) it stays there.
And it stays there.
And it stays there.
And it stays there.
And it stays there.

Essentially unused, pinning down slabs and bucket
spans.

That is, *early* metadata is of very little value, and it
hurts by way of fragmentation and memory occupation.

Metadata these days is compact (into individual
ashift-size objects by lz4, in particular) and for pools
created with spacemap_histogram enabled, it is also
densely packed.   This reduces seek penalties on rotating
media.  Moreover, on non-rotating media, we are even less
in need of holding onto early metadata.

Both ZoL and Solaris Closed ZFS have moved away from
protecting old-but-not-reused metadata altogether.

This is a short step in that direction; some early metadata
that is not reused will be evicted in the first great arc
shrink, whereas before almost none would be.    Metadata
that is reused will survive such shrinks, as will all
metadata (reused or not, independent of age) less than 1/8
of arc_c_min.

As this is runtime tunable, extremely metadata heavy
activity where metadata accesses create significant
seeking delays and each individual metadata record
is *rarely* reused but the metadata working set across a
few txgs is greater than 1/8 of arc_c_min, can set
arc_meta_min as high as they like.

However, this reduction should benefit all users
especially post-ABD.