abd_try_move and arc_abd_move_thread and support
With ABD we still get significant fragmentation visible in
two places: zfs_qcache, which is the parent qcached vmem
arena for ABD chunks and zio and ABD linear allocations;
and in the bucket_65536 arena which services the
overwhelming majority of these quantum cached allocations,
although kstat.spl.misc.spl_misc.spl_buckets_mem_free
reflects our general fragmentation metric.

Like the pure zio_ allocation system, we cannot use a
kmem_move() mechanism for ABD without maintaining a
reverse mapping between each allocated chunk or buffer and
its owner abd.  Unfortunately since abd chunks are small
and therefore much more numerous, this amounts to a
large amount of state that needs to be updated many
times per scattered abd allocation and free (too-large,
too-often - that means bad scalability).

Instead we implement abd_move() and move abds
"asynchronously".

We first introduce locking which allows the memory under
an abd_t to be changed.  While locked, we can copy a
scattered buf's chunks to new chunks and then attach those
chunks to the abd (freeing the old chunks).
Alternatively, for linear abds, we can acquire a new
buffer from the appropriate zio_ allocator, copy the old
contents to it, attach the new buffer to the abd and free
the old buffer.

This "move" mechanism releases "kidnapped" old spans that are
held prisoner by single chunks, which can happen due to
very hot metadata and file data brought very early into
ARC, as kmem_cache (in the absence of any ABD implementation
for kmem_move()) avoids using old kidnapped spans in the
hopes that the kidnapper eventually frees the memory.

Additionally we maintain an age on each abd so that we can
decide that an abd is too young to be worth moving or has
been moved too recently for it to be worthwhile moving again.

ABD *almost* neatly hides the buffers from all
well-behaved clients, except that well-behaved clients are
allowed to use abd_to_buf (and several do!).  Long-lived
inear ABDs are exceedingly rare anyway, so at present we
simply avoid moving linear ABDs (although the mechanism is
present if it becomes useful to move linear ABDs under
some workloads).

In arc.c we start a thread which periodically wakes up and
(if there is apparent fragmentation) spends a few
milliseconds scanning the ARC multilists for candidates
for moving.  Chosen arc_buf_hdr_t candidates that meet
several conditions (those that e.g have an ABD that's not
too young / too recently moved, the arc buffer is not
shared or being written to or presently involved in I/O)
are abd_move()d if they are not linear abds.

We hit a random sublist of each of all four flavours of
ARC multilist during each scan, so long as we have not
exceeded several milliseconds of scanning total.

We hold a multilist sublist lock for no more than two
milliseconds at a time.

While scanning multilist sublists, we hold an ARC hash
lock briefly (copies are almost always very quick).

Technically there is a worst case and EXTREMELY unlikely
wait of a few tens of milliseconds if we have to wait on
xnu for memory and xnu has to wait on paging or
compression to give it to the bucket_##### layer and we
engage with multiple bucket_#####s (e.g. if there is arena
metadata and kmem_alloc activity that has to be done in
order to satisfy an abd chunk allocation)).

However, since each scan is time-limited the result of
holding the hash lock (which we take after acquiring the
multilist sublist lock) is that we will simply go back to
waiting for a second arc_abd_move_thread, so anything
blocked during a long hold will be able to make forward
progress.

Additionally, all of this is at low priority, it's almost
wholly memory rather than compute bound, and offers many
scheduling opportunities for xnu when there is other load
on the machine.  In practice, ZFS patterns of access mean
that sublisting and hashing make lock contention
vanishingly small; additionally, ABD lock contention is
also nearly invisible.