reset the kmem_cache working set data if we move scattered abds
Moving does a lot of allocating, causing full magazines to
be created.

We avoided creating a lot of empty magazines by freeing
directly to the slab layer in the move.

However, arc eviction can create a lot of empty magazines
in the abd cache, since it frees in the normal way (into
to empty magazines where they exist) unless the
kmem_free_to_slab_when_fragmented tunable is 1.

While kmem_update is called every few seconds and will
update the working set, the WS can be too large after we
do a lot of frees, which means reaping may lead to too
many magazines being held in the depot layer rather than
being freed to the slab layer (and then possibly to the
tree of vmem arenas and ultimately xnu).