arc_is_overflowing should return B_FALSE if enough free memory
Technically this makes arc_is_overflowing a lie, because
we can exceed the target, but if we have enough memory
available it shouldn't be a problem.   Additionally,
this reduces the amount of waiting on the reclaim
thread (which can take some time), which should reduce
stalling for a heavy writer (which puts lots of dirty data
into the ARC).