defang cv_wait(&arc_reclaim_waiters_cv,...)
Heavy writers are liable to stall on the cv_wait() in
arc_get_data_impl, and bursty readers and writers are too.

This is most obvious in the case of a system doing
essentially only "tar -xf xcode.tar", which results in a
single-thread mostly heavy and and sometimes bursty
writer, leading to a very slow extraction that gets
especially bursty towards the end.

Break up the cv_wait into a loop that exits in the single
waiting thread case, when arc_is_overflowing returns false
(because either arc is sufficiently close to or below
arc_c, or because there is ample free system memory),
or after a timeout.

Don't send unnecessary cv_signal(&arc_reclaim_thread);

Add some arcstats (and a dmu_tx stat) above msleep() calls
for insight.