zfs.util to return VolumeName when mounted
Following the design of hfs.util, which uses getattrlist to query
VolumeName if device is already mounted.

This corrects Issue #412 in branch issue116_lite, but
make sure datasets are mounted with /dev/disk entries for it
to work.

The changes to zfs_attributes are only cosmetic for the day we implement
getattrlist.