restore from red herring hunt
on zio read we can take undefined scattered buf from
abd_borrow_buf() just fine, because we can overwrite it.
we abd_return_buf_copy() because the read operation should
have overwritten it.

on zio write we want to abd_borrow_buf_copy() because we
want the current abd's contents, and we want to
abd_return_buf() because that will assert that there has
not been a modification of the buffer.