Adapt avg-l's 8372 sync and async zio-s share the same taskq queue
In O3X we don't want to use ZTI_BATCH

We don't reproduce Illumos's ZTI_BATCH semantics, which
sets up a thread which the system makes throughput
guarantees for.   We also can't reproduce ZOL"s appraoch
which is to set up a ZTI_PCT(x) thread which will use up
to x% of a CPU's time.

ZTI_P(a, b) is the most obvious choice, with a and b
initially set to match the values of the other zio_taskqs
entries that use them.

Empirically, with this change we get much higher async
write throughput with slightly lower CPU consumption.