ldi: report long I/O or flush, or overlapping flushes
An I/O ldi_strategy() should not take long to return; if it does,
print a message.

An ldi_ioctl() to do a device cache flush should not take
long to return; if it does, print a message.

Finally, we probably should not have multiple overlapping
ldi_ioctl()s cache flushing to the same device; print a
message when we have more than one overlap.