Exceed zfs_top_maxinflight if scanning has stalled
When a user is logged in at the GUI, scrubs and resilvers
can stall repeatedly for tens of seconds in the loop
waiting for the spa's scrub inflight count to drop below a
threshold derived from zfs_top_maxinflight.

The reason for the stall is unknown. it seems to involve a
wait below IOWorkLoop::closeGate, but not thread
priorities, thread iopolicy, or waiting on device write
cache flushing.  It does not happen in all pools. The
stalls vanish when a userland process creates a
low-bandwidth read workload directly on the /dev/diskXX of
one of the leaf devices in the affected top level
vdev(s).

Additionally, the stalls can be made to vanish by setting
zfs_top_maxinflight very high (e.g. from default 32 to
1024 or 2048)).   However, a very high zfs_top_maxinflight
can introduce enormous latency on any pool that is
scrubbing or resilvering, including pools not affected by
the stall.

This patch replaces the while (maxinflight) cv_wait();
loop with one that will attempt to exit early every
millisecond as long as at it has already been in the loop
for half a second, there has been no other thread
making forward progress on the scrub/resilver, and we
would not simply delay() in the zfs_scan_idle logic
because of recent "important" non-scan I/O.

However, we do not take this early exit if we would have
more than 10 000 scan I/Os in flight (value taken from
spa.c's spa_load_verify_maxinflight).

With this change the scrub/resilver stalls vanish when
looking at "zpool iostat -l pool 1" at least for a pool
comprising rotating disks.  Moreover, latencies do not
increase measurably on that pool.  Latencies on SSD-based
pools also do not increase when scrubbed under this
commit.

* todo:
-more visibility:
-- we should count the "dribbles" in a kstat
-- we should report when we have hit the 10000 ceiling

* also
-further investigation of this stall, and also the
remaining stall when resilvering where neither this
cv_wait nor slowness below vdev_disk_io_start() is
involved.