modify locking in update_pages()
The rw_lock() should serve to serialize whole contending
operations.

Giving up the lock after each page is validated (but not
yet uiomoved) can cause two problems.

The first is that a pagein can happen between the
uiomove() of zfs_write() data into the UBC, which means
that the results of a zfs_vnop_pagein could be *partially*
overwritten by the *older* zfs_write().  The zfs_write()
is therefore not atomic with respect to the UBC cache
update.  A newer zfs_write could already not pass an
in-progress pagein, leading to an asymmetry of atomicity.

(More byzantinely, a pair of sequentially writes competing
against a pagein couuld be scheduled such that some of the
later write's update_pages() pass the update_pages() of
the earlier write; we can't get guarantees from xnu about
that.  Completing the whole update_pages() while locked
prevents this.)

The second (speculatively) is that a pagein could in
principle happen at any time after each ubc page is
validated, including during the uiomove itself.  This
could produce a page with unknown contents taken from the
pagein and write.  It is not clear that Mach's vm map busy
flagging prevents this, or that this is a realistic
scenario.  However, mutexing the whole update_pages()
operation also elminates this (speculative) race.

Finally, the operations in update_pages() are likely to be
very quick, so additional contention from holding the
write lock throughout the whole loop should be minimal.
(We are already paying this penalty in zfs_vnop_pagein().)

Also fix some indentation.