use fewer large (> 1MiB) zio kmem_caches
There is a memory cost to each infrquently used kmem_cache
compared with a memory cost to having kmem_caches that are
large compared to typical allocations.   In the latter
case, long-lived buffers which are small compared to the
cache buffer size are wasteful.   In the former case,
having buf_avail > 0 for long periods of time because of
partial slabs is wasteful.

This is a tricky trade-off for large buf_size kmem_caches,
and we make use several of those (zio_buf for io, mainly)
even after abd scatterization.  Their especially use
increases if one sets
kstat.zfs.darwin.tunable.aggregation_limit
high (e.g. to 16 MiB) which can be a reasonable thing to
do for some workloads (SMR, SSD, numerous raidz
configurations).

We do very few allocations above 1 MiB and few are
long-lived, so the per-cache cost is excessive.  Reduce
the number of kmem_caches above 1 MiB.