mmap fixing: call ubc_msync in sync path; protect critical sections
Once we have mmapped a file we are stuck treating it as if
it is always mmapped until its vnode has been replaced by
the OS.

In particular we have to coordinate with the UBC when it
comes to writes, fsyncs, and msyncs (and an implied msync
on mnomap).  Fsyncs can come from IOCTLs or O_FSYNC, from
the sync property, or from internal zfs activity.   They
should all be ordered with respect to each other; a later
synchronization should not adjust the file or the UBC
before an earlier write/page-dirty/synchronization action.

We have new kstats to monitor this activity, notably in
the pattern kstat.zfs.misc.vnops...want_lock.  Nonzero
values in those represent a serialization of
state-mutation (in the UBC and/or in the zfs dataset) for
a single file.

Some highlights:

cn:

window 1:

./fsx -c 1000 -o 134217728 -F 134217728 -m 0:10000 -S
foobar

window 2:

gdd if=somefile of=foobar conv=notrunc

duelling fsx -c 1000 -o 134217728 -F 134217728 -m 0:10000 foobar

cn+1

convert rw_enter(&zp->z_map_lock) to a spin-sleep

we want to track contention.

this helps decide whether we want to hoist z_map_lock
above the rangelocks, which will require surgery in
zfs_vnop_write similar to zfs_fsync.

this violates DRY a bit; perhaps this should be
genericized along the lines of zfs_fsync_rw_lock() (or all
those utility funcs should be genericized into something
inline-able).

cn+2

Use a CAS spin-sleep loop to serialize in zfs_fsync

Multiple zfs_fsync may be in flight for a given file.

There is complex mutex locking below zfs_fsync, and
unfortunately zfs_fsync may also be called recursively on
the same file if it is also mmapped.

We can build an in-function serializing lock by using
an extra znode_t field and an atomic compare-and-swap.