restructuring in update_pages
We should not take z_map_lock if z_is_mapped is false.

If z_is_mapped is true, it will remain true for the
lifetime of the znode.   But if it is false, another
thread might mmap the file, turning it true.   So we if
z_is_mapped is false, we hold z_lock.