don't lock in mnomap, PUSHALL for mapped files
we don't take a range lock in mnomap anyway, although
perhaps we should.   if there are pages to go out, they
will hit pageoutv2, which will do RL->z_map_lock in the
right order, so this is likely better.

also, be slightly more nfs-like in terms of use of
UBC_PUSHALL vs UBC_PUSHDIRTY.   ALL when the file is
mapped, and SYNC on mnomap.