zfs_vnop_advlock(): a small shim around (spl_)lf_advlock
We hand off to spl_lf_advlock unless spl_ubc_is_mapped(vp,
NULL) is true, in which case we return EAGAIN.

     The lockf() function will fail if:

     [EAGAIN]           The argument function is F_TLOCK or F_TEST and the section is already
                        locked by another process.

which is not quite right, but it's what Illumos's
zfs_frlock() returns.

lf_setlock() [called by lf_advlock] can return EAGAIN if
the lock is noonblocking.

(It also can return EDEADLK.   Mostly it is expected to return "???" :^) )