avoid panic due to automatic freeing of UPLs with no valid pages left
While scanning from the front of the UPL and having done
at least one _FREE_ON_EMPTY commit or abort, under memory
pressure the VM will destroy the UPL if there are only
non-present pages remaining.  When that happens, aborting
the remaining range results in KERN_FAILURE (error 5), and
the subsequent ubc_upl_unmap is liable to provoke a kernel
panic.

Once the UPL is created, scan from the back of the UPL
until we find a valid page.   If we find none, we're done.
If we find one, then focus on only the pages between the
start of the UPL and that valid page; once it is committed
or aborted, the UPL is liable to be automatically
destroyed, so we need to make sure that we unmap before
disposing of that page.    (We were doing that already for
the final page of the UPL, but if it was absent and the
kernel destroyed the UPL before we got to it, we lost).