be forgiving of sub-page underlocking, and pageoutv2 printf->dprintf
the risk with sub-page underlocking in
pagein/pageout (which happens when they recover the TSD
from rl) is that we fall through and hit a mutex below
memory_object_lock_request.

pagein is always page-sized and page-aligned, so will
collide with an underlocked earlier pageout, or block a
later underlocked pageout, where there is overlap.

pageouts that do not come with a range lock acquired by
the caller (e.g. pageouts driven by the kernel itself)
will come in pages, so again the underlocking is OK.

finally there is the case where two pageouts into the same
page happen; we are protected by z_map_lock in this case.

(also it would not be awful to hit
memory_object_lock_request's lock in these cases, since
the parallel pageouts should not prevent each others'
progress at that point))