range locking when VM gives us a different range
for some reason, the VM can give us pages outside the range of the
ubc_msync call.  possibly these are precious pages (?).

try to lock the bit not covered by the RL we inherited when this is the
case.  (if we can't, we race, which means we can crash into another
thread's range-lock, or there can be a crash in
memory_object_lock_request ).