replace cluster_copy_ubc_data in mappedread_new
cluster_copy_ubc_data has a habit of getting stuck in
memory_control_object_uiomove, presumably because a page
in the UBC is dirty or has gone missing, requiring
resp. either a pageout or a pagein.

although creating a upl and mapping it is strictly less
efficient than using cluster_copy_ubc_data (which avoids
non-physmapping), we can at least inspect the page state
and complain in these cases, before actually accessing the
pages (and possibly getting stuck as a result).

at the moment, if a dirty page is encountered, note it via
printf, but otherwise treat it as any other valid page.

an invalid page (e.g. one that is has the absent bit set,
or has no phyical address) will error out.