slight changes to zfs_vfs_sync & callback
don't zil_commit on MNT_NOWAIT, which is issued from time
to time

VNODE_RETURNED in the MNT_NOWAIT case where the file has
never been through zfs_ubc_msync.  We don't want to be the
first to sync a file that has never been the subject of a
zfs_write, because we don't want to collide with an
msync(2) over which we have no control and with which we
race into virtual memory locks, risking deadlocks.  We
should VNODE_CLAIMED in the MNT_WAIT/MNT_DWAIT case to
have the vnode_iterate stick around for a while, but in
their absence, we should just stop bothering.