adjust z_mr_sync for locking out fsync in light of write changes
The general idea is that a file which has been disturbed
in any way should not be subjected to zfs_vfs_fsync; we
also want to blockade the fsync vnop and fcntl, which can
come at inopportune times, e.g. exposing the (former?)
non-atomicity of the zfs_write() sync_range implementation.