(shrink) memory object lock request hang below zfs_znode_getvnode
do not ubc_setsize(zp->z_size) in zfs_znode_getvnode if
ubc_getsize is in the same page as zp->z_size.

this will overwhelmingly affect files that are no more
than 4095 bytes long, although we need to do it because of
the small risk of deadlocking below ubc_setsize.