reduce memory/pager burden of fill_holes
only fill pages which are not present, that is, which do
not have a non-null physaddr.    pages which are not valid
can be present but marked other-than-absent (i.e., while
the absent bit cannot be set, the speculative, dirty, or
pageout bits may be set).   writing over them causes a
COW, increasing memory occupancy, and causing races down
through pagein (zfs_read->uiomove64->bcopy->...->kernel_trap->
...->vm_fault_page->vnode_pager_data_request->vnode_pager_cluster_read->
vnode_pagein->zfs_vnop_pagein.

at best the pagein is simply overwritten with identical
data in fill_holes.  at worst, the page is marked pageout
and is in pageout when pagein blithely decides that it is
probably pageout requesting the page, so it is safe to
read it from the dmu.  result: vm memory recovery
pageout (triggered by non-xnu memory processing,
noticeably when in low memory and chrome (or often some
sqlite-using process) is doing busy read i/o) has its
in-process-of-paging-out page replaced by a page from
zfs_read.  what is stored in the pool is correct, but what
is held in memory is stale: application notices and fails.